typing: Start mypy setup
Add mypy and configure its run on the project. For now is just a
placeholder and *will not* move forward because mypy is full of errors,
and it's reporting lots of false positives.

Continuing in this direction, by ignoring the cases where mypy fails to
correctly report, errors, will just pollute the code.

Waiting until a new version addresses these issues

TL;DR: mypy sucks (at least for now).