typing: Ignore typehinting
mypy is right in this check because both concatenation and the
``startswith`` method work with bytes and string, so long as both parts
are of the same kind (both strings or both bytes). If they are mixed
(for example ``<string>.startswith(<byte>)`` or ``<string> + <byte>``),
then it *will* fail.

However, the code makes this assertion right before, so that case can
never be reached.

That's why it's safe to ignore.