Don't hold extra properties
Those attributes that are only required for the output file, can be (and
are) removed, since the PyCompressor object only needs to work with the
destination, that can be computed at initialization time.

This also makes the object complete at initialization time, it will
contain everything it needs to work right after its creation.